Make a separate module for text_edit
